moved `compute_f1_score` to static_funcs.py
